Split mobile attach target type
